30/5/2019 4:16PM add collisionTime, double dt to OnCollision

+++ LearningDirectX2/BigBullet.cpp
@@ -25,7 +25,7 @@ BigBullet::BigBullet() {
 BigBullet::~BigBullet() {
 }
 
-void BigBullet::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, float dt) {
+void BigBullet::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, double dt) {
 }
 
 void BigBullet::Update(double dt) {
