+++ LearningDirectX2/Sparta.h
@@ -8,7 +8,7 @@ class Sparta : public Enemy {
 public:
 	Sparta();
 	~Sparta();
-	virtual void OnCollision(Entity *impactor, Entity::SideCollision side, float collisionTime, float dt = 1.0/60);
+	virtual void OnCollision(Entity *impactor, Entity::SideCollision side, float collisionTime, double dt = 1.0/60);
 	virtual void Update(double dt);
 	//virtual float GetWidth();
 	//virtual float GetHeight();
